모든 팔로워들에게 알림보내기 batchInsert로 변경

+++ src/main/java/com/brandpark/sharemusic/api/v1/album/dto/AlbumInfoResponse.java
@@ -11,7 +11,7 @@ public class AlbumInfoResponse {
     private String title;
     private String albumImage;
     private String description;
-    private int trackCount;
+    private Long trackCount;
     private List<TrackInfoResponse> tracks;
     private Long accountId;
     private LocalDateTime createdDate;
